backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.32
psycopg[binary]==3.2.1
alembic==1.13.2
pydantic==2.8.2
pydantic-settings==2.4.0
python-multipart==0.0.9
python-dotenv==1.0.1
passlib==1.7.4
bcrypt==3.2.2
PyJWT==2.9.0
